+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## staged
 
+- Added FOT linear formulation in rectangular coordinate frame
 - Fixed lower voltage magnitude limits for FBS linear formulation
 - Fix bug where the lower and upper bound variables created in `variable_mx_hermitian` can be the wrong type
 - Fix bug in correct_bus_types! and eng2math functions where reference bus was being overwritten depending on the order generation objects were parsed